Add ocamlformat formatter.
We have to handle a case when buffer without filename is formatted, as
ocamlformat expects `--name`, `--impl`, or `--intf` to be provided. In
case `--name` is provided, ocamlformat will use path of that file as
starting point for searching for .ocamlformat. When formatting buffers
with no filename, we provide `--impl` explicitly, to indicate that
ocamlformat should parse the input as an implementation rather than an
interface.